fixed the bitmap loader. Implemented the new bitmap loader into X11 and also implemented it into SDL (we no longer rely on SDL's bitmap loader). fixed a few bugs also.

+++ src/extlib/none.c
@@ -65,7 +65,7 @@ Lib_SetColorKey(v_object *vobj, u32 key)
 }
 
 void
-Lib_SetAlpha(v_object *vobj, u32 alpha)
+Lib_SetAlpha(v_object *vobj, u8 alpha)
 {
 	
 }
